requirements.txt
Flask==3.1.2
GitPython==3.1.45
jedi==0.19.2
langchain_community==0.3.29
langgraph==0.6.7
libclang==18.1.1
litellm==1.77.1
lxml==6.0.2
numpy==2.2.6
ollama==0.5.3
prettytable==3.16.0
pyelftools==0.32
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
rustworkx==0.17.1
unidiff==0.7.5
